refactor(tui): split tui.rs into multiple modules for better maintainability
- Created src/tui/ directory structure with 5 modules:
  - focus.rs: Focus management (100% coverage)
  - app.rs: App state and business logic (97.72% coverage)
  - event.rs: Event handling (93.81% coverage)
  - ui.rs: UI rendering (not directly testable)
  - runner.rs: TUI entry point (not directly testable)
- Migrated 20 existing tests to app.rs
- Added 32 new tests for focus.rs and event.rs
- Added 7 edge case tests to app.rs
- All testable modules now exceed 85% code coverage
- Total test count: 142 tests passing
